% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessCheck.r
\docType{data}
\name{AccessCheck}
\alias{AccessCheck}
\title{AccessCheck Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AccessCheck
}
\description{
AccessCheck Class
}
\section{Fields}{

\describe{
\item{\code{aclKey}}{}

\item{\code{permissions}}{}
}}

\keyword{datasets}
