% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ace.r
\docType{data}
\name{Ace}
\alias{Ace}
\title{Ace Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Ace
}
\description{
Ace Class
}
\section{Fields}{

\describe{
\item{\code{principal}}{}

\item{\code{permissions}}{}
}}

\keyword{datasets}
