% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acl.r
\docType{data}
\name{Acl}
\alias{Acl}
\title{Acl Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Acl
}
\description{
Acl Class
}
\section{Fields}{

\describe{
\item{\code{aclKey}}{}

\item{\code{aces}}{}
}}

\keyword{datasets}
