% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AclData.r
\docType{data}
\name{AclData}
\alias{AclData}
\title{AclData Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AclData
}
\description{
AclData Class
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{acl}}{}
}}

\keyword{datasets}
