% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedSearch.r
\docType{data}
\name{AdvancedSearch}
\alias{AdvancedSearch}
\title{AdvancedSearch Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AdvancedSearch
}
\description{
AdvancedSearch Class
}
\section{Fields}{

\describe{
\item{\code{searchFields}}{}

\item{\code{start}}{}

\item{\code{maxHits}}{}
}}

\keyword{datasets}
