% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiClient.r
\docType{data}
\name{ApiClient}
\alias{ApiClient}
\title{ApiClient Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ApiClient
}
\description{
Generic API client for OpenAPI client library builds.
OpenAPI generic API client. This client handles the client-
server communication, and is invariant across implementations. Specifics of
the methods and models for each application are generated from the OpenAPI Generator
templates.
}
\details{
NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
Ref: https://openapi-generator.tech
Do not edit the class manually.
}
\keyword{datasets}
