% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association.r
\docType{data}
\name{Association}
\alias{Association}
\title{Association Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Association
}
\description{
Association Class
}
\section{Fields}{

\describe{
\item{\code{key}}{}

\item{\code{src}}{}

\item{\code{dst}}{}

\item{\code{details}}{}
}}

\keyword{datasets}
