% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationType.r
\docType{data}
\name{AssociationType}
\alias{AssociationType}
\title{AssociationType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AssociationType
}
\description{
AssociationType Class
}
\section{Fields}{

\describe{
\item{\code{entityType}}{}

\item{\code{src}}{}

\item{\code{dst}}{}

\item{\code{bidirectional}}{}
}}

\keyword{datasets}
