% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auth0userBasic.r
\docType{data}
\name{Auth0userBasic}
\alias{Auth0userBasic}
\title{Auth0userBasic Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Auth0userBasic
}
\description{
Auth0userBasic Class
}
\section{Fields}{

\describe{
\item{\code{user_id}}{}

\item{\code{email}}{}

\item{\code{nickname}}{}

\item{\code{username}}{}

\item{\code{roles}}{}

\item{\code{com/openlattice/organization}}{}
}}

\keyword{datasets}
