% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Authorization.r
\docType{data}
\name{Authorization}
\alias{Authorization}
\title{Authorization Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Authorization
}
\description{
Authorization Class
}
\section{Fields}{

\describe{
\item{\code{aclKey}}{}

\item{\code{permissions}}{}
}}

\keyword{datasets}
