% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthorizationsApi.r
\docType{data}
\name{AuthorizationsApi}
\alias{AuthorizationsApi}
\title{Authorizations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthorizationsApi
}
\description{
openlattice.Authorizations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

check_authorizations Check authorizations


get_accessible_objects Returns paged results for all authorized objects of specified objectType, that the current user has specified permission for.

}
}

\keyword{datasets}
