% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthorizedObjectsSearchResult.r
\docType{data}
\name{AuthorizedObjectsSearchResult}
\alias{AuthorizedObjectsSearchResult}
\title{AuthorizedObjectsSearchResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthorizedObjectsSearchResult
}
\description{
AuthorizedObjectsSearchResult Class
}
\section{Fields}{

\describe{
\item{\code{pagingToken}}{}

\item{\code{authorizedObjects}}{}
}}

\keyword{datasets}
