% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkDataCreation.r
\docType{data}
\name{BulkDataCreation}
\alias{BulkDataCreation}
\title{BulkDataCreation Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BulkDataCreation
}
\description{
BulkDataCreation Class
}
\section{Fields}{

\describe{
\item{\code{entities}}{}

\item{\code{associations}}{}
}}

\keyword{datasets}
