% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataApi.r
\docType{data}
\name{DataApi}
\alias{DataApi}
\title{Data operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataApi
}
\description{
openlattice.Data
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

create_associations Creates a new set of associations.


delete_all_entities_from_entity_set Clears the Entity matching the given Entity id and all of its neighbor Entities


delete_entities Deletes multiple entities from an entity set.


get_entity_set_size Gets the number of entities in an entity set.


load_entity_set_data Gets an iterable containing the entity data, using property type FQNs as key


load_filtered_entity_set_data Gets a list of entities by UUIDs

}
}

\keyword{datasets}
