% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEdge.r
\docType{data}
\name{DataEdge}
\alias{DataEdge}
\title{DataEdge Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataEdge
}
\description{
DataEdge Class
}
\section{Fields}{

\describe{
\item{\code{src}}{}

\item{\code{dst}}{}

\item{\code{data}}{}
}}

\keyword{datasets}
