% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIntegrationsApi.r
\docType{data}
\name{DataIntegrationsApi}
\alias{DataIntegrationsApi}
\title{DataIntegrations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataIntegrationsApi
}
\description{
openlattice.DataIntegrations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_entity_key_ids Get entity key IDs


integrate_entity_and_association_data Integrate entity and association data

}
}

\keyword{datasets}
