% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSearchResult.r
\docType{data}
\name{DataSearchResult}
\alias{DataSearchResult}
\title{DataSearchResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSearchResult
}
\description{
DataSearchResult Class
}
\section{Fields}{

\describe{
\item{\code{numHits}}{}

\item{\code{hits}}{}
}}

\keyword{datasets}
