% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDM.r
\docType{data}
\name{EDM}
\alias{EDM}
\title{EDM Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDM
}
\description{
EDM Class
}
\section{Fields}{

\describe{
\item{\code{namespaces}}{}

\item{\code{schemas}}{}

\item{\code{propertyTypes}}{}

\item{\code{entityTypes}}{}

\item{\code{associationTypes}}{}
}}

\keyword{datasets}
