% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdmApi.r
\docType{data}
\name{EdmApi}
\alias{EdmApi}
\title{Edm operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EdmApi
}
\description{
openlattice.Edm
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_dst_entity_type_to_association_type Update the AssociationType dst entity types for the given AssociationType UUID by adding the given EntityType UUID.


add_property_type_to_entity_type Updates the EntityType definition for the given EntityType UUID by adding the given PropertyType UUID.


add_src_entity_type_to_association_type Update the AssociationType src entity types for the given AssociationType UUID by adding the given EntityType UUID.


create_association_type Creates a new AssociationType definition, if it doesn\&quot;t exist.


create_empty_schema Creates an empty schema, if it doesn&#39;t exist. If schema exists then no action is taken.


create_entity_sets Create new EntitySet definitions if they don\&quot;t exist.


create_entity_type Creates a new EntityType definition, if it doesn&#39;t exist.


create_property_type Creates a new PropertyType definition, if it doesn\&quot;t exist.


create_schema_if_not_exists Creates an empty schema, if it doesn&#39;t exist. If schema exists then no action is taken.


delete_association_type Delete the AssociationType definition for the given AssociationType UUID.


delete_entity_set Deletes the EntitySet definition for the given EntitySet UUID.


delete_entity_type Deletes the EntityType definition for the given EntityType UUID.


delete_property_type Deletes the PropertyType definition for the given PropertyType UUID.


force_delete_property_type Deletes the PropertyType definition for the given PropertyType UUID regardless of whether or not there is data associated with it.


force_remove_property_type_from_entity_type Updates the EntityType definition for the given EntityType UUID by removing the given PropertyType UUID, regardless of whether or not there is data associated with the entity type.


get_all_association_entity_types Gets all association EntityType definitions.


get_all_available_association_types Get all available associations for the given AssociationType UUID.


get_all_entity_set_property_metadata Get all property type metadata for an entity set.


get_all_entity_sets Get all EntitySet definitions.


get_all_entity_type_property_metadata Get all EntityType propertyType metadata


get_all_entity_types Gets all EntityType definitions.


get_all_property_types Gets all PropertyType definitions.


get_all_property_types_in_namespace Gets all PropertyType definitions under the given namespace.


get_all_property_usage_summaries Get Property Usage Summary for all properties.


get_association_type Get the AssociationType definition for the given AssociationType UUID.


get_association_type_details Get details about the AssociationType for the given AssociationType UUID.


get_entity_data_model Gets the entity data model, including namespaces, schemas, entity types, association types, and property types.


get_entity_data_model_diff Gets the changes between the existing entity data model and the entity data model passed in, including schemas, association types, entity types, and property types.


get_entity_set Get the EntitySet definition for the given EntitySet UUID.


get_entity_set_id Gets the EntitySet UUID for the given EntitySet name.


get_entity_set_property_metadata Get specified property type metadata for an entity set.


get_entity_type Gets the EntityType definition for the given EntityType UUID.


get_entity_type_hierarchy Get the EntityType hierarchy for the given EntityType UUID.


get_entity_type_id Gets the EntityType UUID for the given EntityType FQN.


get_entity_type_property_metadata Get EntityType propertyType metadata


get_property_type Gets the PropertyType definition for the given PropertyType UUID.


get_property_type_id Gets the PropertyType UUID for the given PropertyType FQN.


get_property_usage_summary Get Property Usage Summary for property with given ID.


get_schema_contents Gets the schema contents for a corresponding namespace and name.


get_schemas Gets all schemas.


get_schemas_in_namespace Gets all schemas associated with a given namespace and accessible by the caller.


remove_dst_entity_type_from_association_type Updates the AssociationType dst entity types for the given AssociationType UUID by removing the given EntityType UUID.


remove_property_type_from_entity_type Updates the EntityType definition for the given EntityType UUID by removing the given PropertyType UUID.


remove_src_entity_type_from_association_type Updates the AssociationType src entity types for the given AssociationType UUID by removing the given EntityType UUID.


reorder_property_types_in_entity_type Updates the EntityType definition for the given EntityType UUID by reordering its properties as specified by the provided list.


update_entity_data_model Updates the entity data model, including schemas, entity types, association types, and property types.


update_entity_set_meta_data Updates the EntityType definition for the given EntitySet UUID with the given metadata.


update_entity_set_property_metadata Updates the property type metadata for the given entity set.


update_entity_type_meta_data Updates the EntityType definition for the given EntityType UUID with the given metadata.


update_entity_type_property_metadata Update EntityType Property metadata


update_property_type_meta_data Updates the PropertyType definition for the given PropertyType UUID with the given metadata.


update_schema Edits the schema contents for a corresponding namespace and name.

}
}

\keyword{datasets}
