% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMdiff.r
\docType{data}
\name{EDMdiff}
\alias{EDMdiff}
\title{EDMdiff Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EDMdiff
}
\description{
EDMdiff Class
}
\section{Fields}{

\describe{
\item{\code{conflicts}}{}

\item{\code{diff}}{}
}}

\keyword{datasets}
