% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdmRequest.r
\docType{data}
\name{EdmRequest}
\alias{EdmRequest}
\title{EdmRequest Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EdmRequest
}
\description{
EdmRequest Class
}
\section{Fields}{

\describe{
\item{\code{action}}{}

\item{\code{propertyTypes}}{}

\item{\code{entityTypes}}{}
}}

\keyword{datasets}
