% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entity.r
\docType{data}
\name{Entity}
\alias{Entity}
\title{Entity Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Entity
}
\description{
Entity Class
}
\section{Fields}{

\describe{
\item{\code{key}}{}

\item{\code{details}}{}
}}

\keyword{datasets}
