% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityKey.r
\docType{data}
\name{EntityKey}
\alias{EntityKey}
\title{EntityKey Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityKey
}
\description{
EntityKey Class
}
\section{Fields}{

\describe{
\item{\code{entitySetId}}{}

\item{\code{entityId}}{}
}}

\keyword{datasets}
