% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntitySet.r
\docType{data}
\name{EntitySet}
\alias{EntitySet}
\title{EntitySet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntitySet
}
\description{
EntitySet Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{id}}{}

\item{\code{title}}{}

\item{\code{contacts}}{}

\item{\code{entityTypeId}}{}

\item{\code{description}}{}

\item{\code{linking}}{}

\item{\code{linked_entity_sets}}{}

\item{\code{external}}{}

\item{\code{organizationId}}{}
}}

\keyword{datasets}
