% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntitySetSelection.r
\docType{data}
\name{EntitySetSelection}
\alias{EntitySetSelection}
\title{EntitySetSelection Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntitySetSelection
}
\description{
EntitySetSelection Class
}
\section{Fields}{

\describe{
\item{\code{ids}}{}

\item{\code{properties}}{}
}}

\keyword{datasets}
