% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityType.r
\docType{data}
\name{EntityType}
\alias{EntityType}
\title{EntityType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityType
}
\description{
EntityType Class
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{schemas}}{}

\item{\code{key}}{}

\item{\code{properties}}{}

\item{\code{propertyTags}}{}

\item{\code{basetype}}{}

\item{\code{category}}{}
}}

\keyword{datasets}
