% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityTypePropertyMetadata.r
\docType{data}
\name{EntityTypePropertyMetadata}
\alias{EntityTypePropertyMetadata}
\title{EntityTypePropertyMetadata Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityTypePropertyMetadata
}
\description{
EntityTypePropertyMetadata Class
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{description}}{}
}}

\keyword{datasets}
