% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullQualifiedName.r
\docType{data}
\name{FullQualifiedName}
\alias{FullQualifiedName}
\title{FullQualifiedName Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FullQualifiedName
}
\description{
FullQualifiedName Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{namespace}}{}
}}

\keyword{datasets}
