% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrationResults.r
\docType{data}
\name{IntegrationResults}
\alias{IntegrationResults}
\title{IntegrationResults Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IntegrationResults
}
\description{
IntegrationResults Class
}
\section{Fields}{

\describe{
\item{\code{entityCount}}{}

\item{\code{associationCount}}{}

\item{\code{associationsEntityKeyMappings}}{}

\item{\code{entitySetsEntityKeyMappings}}{}
}}

\keyword{datasets}
