% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaterializedViewAccount.r
\docType{data}
\name{MaterializedViewAccount}
\alias{MaterializedViewAccount}
\title{MaterializedViewAccount Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MaterializedViewAccount
}
\description{
MaterializedViewAccount Class
}
\section{Fields}{

\describe{
\item{\code{username}}{}

\item{\code{credential}}{}
}}

\keyword{datasets}
