% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaDataUpdate.r
\docType{data}
\name{MetaDataUpdate}
\alias{MetaDataUpdate}
\title{MetaDataUpdate Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MetaDataUpdate
}
\description{
MetaDataUpdate Class
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{title}}{}

\item{\code{contacts}}{}

\item{\code{description}}{}

\item{\code{type}}{}

\item{\code{linking}}{}

\item{\code{linked_entity_sets}}{}

\item{\code{external}}{}

\item{\code{pii}}{}

\item{\code{defaultShow}}{}

\item{\code{propertyTags}}{}

\item{\code{organizationId}}{}
}}

\keyword{datasets}
