% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborEntityDetails.r
\docType{data}
\name{NeighborEntityDetails}
\alias{NeighborEntityDetails}
\title{NeighborEntityDetails Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NeighborEntityDetails
}
\description{
NeighborEntityDetails Class
}
\section{Fields}{

\describe{
\item{\code{associationEntitySet}}{}

\item{\code{associationDetails}}{}

\item{\code{neighborEntitySet}}{}

\item{\code{neighborId}}{}

\item{\code{neighborDetails}}{}
}}

\keyword{datasets}
