% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborEntityIds.r
\docType{data}
\name{NeighborEntityIds}
\alias{NeighborEntityIds}
\title{NeighborEntityIds Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NeighborEntityIds
}
\description{
NeighborEntityIds Class
}
\section{Fields}{

\describe{
\item{\code{associationId}}{}

\item{\code{neighborId}}{}

\item{\code{src}}{}
}}

\keyword{datasets}
