% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborSearchFilter.r
\docType{data}
\name{NeighborSearchFilter}
\alias{NeighborSearchFilter}
\title{NeighborSearchFilter Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NeighborSearchFilter
}
\description{
NeighborSearchFilter Class
}
\section{Fields}{

\describe{
\item{\code{entityKeyIds}}{}

\item{\code{src}}{}

\item{\code{dst}}{}

\item{\code{edge}}{}
}}

\keyword{datasets}
