% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizationMember.r
\docType{data}
\name{OrganizationMember}
\alias{OrganizationMember}
\title{OrganizationMember Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OrganizationMember
}
\description{
OrganizationMember Class
}
\section{Fields}{

\describe{
\item{\code{principal}}{}

\item{\code{profile}}{}

\item{\code{roles}}{}
}}

\keyword{datasets}
