% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganizationsApi.r
\docType{data}
\name{OrganizationsApi}
\alias{OrganizationsApi}
\title{Organizations operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OrganizationsApi
}
\description{
openlattice.Organizations
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

add_member Add member to an organization


add_role_to_user Add a role to a user


assemble_entity_sets Materializes entity sets into the organization database.


create_organization_if_not_exists Create an organisation if it doesn&#39;t exist.


create_role Create role


delete_role Remove role for an organization


destroy_organization Remove an organisation from the organizationId


get_all_users_of_role Get members of a role for an organization from a roleId


get_auto_approved_email_domains Get auto-approved email domains


get_flagged_organization_entity_sets Get the entity sets for an organization for a certain flag


get_members Get members of a certain organization


get_organization Get an organisation from the organizationId


get_organization_entity_sets Get the entity sets for an organization for a certain filter


get_organizations Get all organisations


get_role Get role for an organization from a roleId


get_roles Get roles for an organization


remove_auto_approved_email_domains Remove auto-approved email domains


remove_member Remove member from an organization


remove_role_from_user Remove a role from a user


set_auto_approved_email_domain Set auto-approved email domains


update_description Update the organisation description


update_role_description Update role description for an organization from a roleId


update_role_title Update role title for an organization from a roleId


update_title Update the organisation title

}
}

\keyword{datasets}
