% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Principal.r
\docType{data}
\name{Principal}
\alias{Principal}
\title{Principal Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Principal
}
\description{
Principal Class
}
\section{Fields}{

\describe{
\item{\code{type}}{}

\item{\code{id}}{}
}}

\keyword{datasets}
