% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrincipalApi.r
\docType{data}
\name{PrincipalApi}
\alias{PrincipalApi}
\title{Principal operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrincipalApi
}
\description{
openlattice.Principal
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_all_users Get all users.


get_current_roles Get current roles.


get_materialized_view_account Get the db access credentials.


get_user Get the user for the given id.


search_all_users_by_email Get the user id for the given email address.

}
}

\keyword{datasets}
