% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyType.r
\docType{data}
\name{PropertyType}
\alias{PropertyType}
\title{PropertyType Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PropertyType
}
\description{
PropertyType Class
}
\section{Fields}{

\describe{
\item{\code{title}}{}

\item{\code{id}}{}

\item{\code{type}}{}

\item{\code{description}}{}

\item{\code{schemas}}{}

\item{\code{datatype}}{}

\item{\code{piiField}}{}

\item{\code{multiValued}}{}

\item{\code{analyzer}}{}
}}

\keyword{datasets}
