% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyUsageSummary.r
\docType{data}
\name{PropertyUsageSummary}
\alias{PropertyUsageSummary}
\title{PropertyUsageSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PropertyUsageSummary
}
\description{
PropertyUsageSummary Class
}
\section{Fields}{

\describe{
\item{\code{entityTypeId}}{}

\item{\code{entitySetId}}{}

\item{\code{entitySetName}}{}

\item{\code{count}}{}
}}

\keyword{datasets}
