% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Role.r
\docType{data}
\name{Role}
\alias{Role}
\title{Role Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Role
}
\description{
Role Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{organizationId}}{}

\item{\code{principal}}{}

\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{@class}}{}
}}

\keyword{datasets}
