% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schema.r
\docType{data}
\name{Schema}
\alias{Schema}
\title{Schema Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Schema
}
\description{
Schema Class
}
\section{Fields}{

\describe{
\item{\code{entityTypes}}{}

\item{\code{propertyTypes}}{}

\item{\code{fqn}}{}
}}

\keyword{datasets}
