% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchApi.r
\docType{data}
\name{SearchApi}
\alias{SearchApi}
\title{Search operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SearchApi
}
\description{
openlattice.Search
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

execute_advanced_entity_set_data_query Executes a search over the data of a given entity set to find rows that match the search term


execute_entity_neighbor_search Executes a search for all neighbors of an entity that are connected by an association


execute_filtered_entity_neighbor_id_search Executes a search for all neighbors of multiple entities of the same entity set that are connected by an association and returns a simple version of the neighborDetails


execute_filtered_entity_neighbor_search Executes a search for all neighbors of multiple entities of the same entity set that are connected by an association

}
}

\keyword{datasets}
