% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchDetails.r
\docType{data}
\name{SearchDetails}
\alias{SearchDetails}
\title{SearchDetails Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SearchDetails
}
\description{
SearchDetails Class
}
\section{Fields}{

\describe{
\item{\code{searchTerm}}{}

\item{\code{property}}{}

\item{\code{exact}}{}
}}

\keyword{datasets}
