% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecurablePrincipal.r
\docType{data}
\name{SecurablePrincipal}
\alias{SecurablePrincipal}
\title{SecurablePrincipal Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SecurablePrincipal
}
\description{
SecurablePrincipal Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{principal}}{}

\item{\code{title}}{}

\item{\code{description}}{}
}}

\keyword{datasets}
